Add "sopelshh", for multiplexing Bucket & Sopel
Sometimes the best way to get all the functionality you want in an IRC
bot is to just run two (or more) bots behind the same nick using a
bouncer. Bucket sometimes responds to the Sopel bot's commands, which
can be confusing, but this plugin tells Bucket to stay quiet unless the
command exactly matches a factoid.